+++ CHANGELOG.md
@@ -12,6 +12,7 @@
 - Support markdown extensions to load in config.toml under `[markdown]` section with key `extensions`
 - Set URL for `posts.all_posts` in config.toml under `[posts]` section with key `base_url`.
 - Add `static/htmd.css` and `static/htmd.js` to add a fragment link next to headings
+- Add search with [Pagefind](https://pagefind.app/) to `htmd build`
 ### Changed
 - Improve Atom feed
     - Use fully qualified URLs